refactor: remove Head component and migrate to app router metadata
