+++ src/app/categorias-naturaleza/page.jsx
@@ -9,7 +9,7 @@ import LoadingSpinner from "@/components/Loading/LoadingSpinner";
 import DepartamentoCard from "@/components/Card/DepartamentoCard";
 import Pagination from "@/components/ui/Pagination";
 import PageSizeSelector from "@/components/ui/PageSizeSelector";
-import Head from "next/head";
+
 
 export default function CategoriasNaturaleza() {
   const pageTitle = metadata.catNat.title;
@@ -50,9 +50,7 @@ export default function CategoriasNaturaleza() {
 
   return (
     <>
-      <Head>
-        <title>{`${pageTitle} • Colombia 360`}</title>
-      </Head>
+      
       <PageSizeSelector pageSize={pageSize} setPageSize={setPageSize} />
       <PageSection
         title={pageTitle}
