+++ src/app/prueba/page.jsx
@@ -5,7 +5,7 @@ import "./mapas.css";
 
 import { ColombiaMap } from "@/components/Map/ColombiaMap";
 import PageSection from "@/components/PageSection";
-import Head from "next/head";
+
 
 import { metadata } from "@/components/metadata";
 
@@ -14,9 +14,7 @@ export default function Prueba() {
 
   return (
     <>
-      <Head>
-        <title>{`${pageTitle} • Colombia 360`}</title>
-      </Head>
+      
       <PageSection title={pageTitle}>
         <section className="flex w-full items-center justify-center rounded-lg bg-slate-950/90">
           <div
